+++ exercises/object-oriented-programming/src/main/java/com/cbfacademy/accounts/SavingsAccount.java
@@ -17,7 +17,7 @@ public class SavingsAccount extends Account {
      * @param interestRate The interest rate associated with the savings account, represented as a decimal value.
      *                     For example, an interest rate of 0.05 represents 5% interest.
      */
-    public SavingsAccount(long accountNumber, double balance, double interestRate) {
+    public SavingsAccount(int accountNumber, double balance, double interestRate) {
         super(accountNumber, balance);
         this.interestRate = interestRate;
     }
